requirements.txt
dacite>=1.8.0
Requests>=2.31.0
setuptools>=59.6.0
python-dotenv